+++ core/embed/rust/src/ui/layout_eckhart/ui_firmware.rs
@@ -783,7 +783,7 @@ impl FirmwareUI for UIEckhart {
         Ok(layout)
     }
 
-    fn show_device_menu(failed_backup: bool, battery_percentage: usize, paired_devices: Vec<TString<'static>, 1>) -> Result<impl LayoutMaybeTrace, Error> {
+    fn show_device_menu(failed_backup: bool, battery_percentage: u8, paired_devices: Vec<TString<'static>, 1>) -> Result<impl LayoutMaybeTrace, Error> {
         let layout = RootComponent::new(DeviceMenuScreen::new(failed_backup, battery_percentage, paired_devices));
         Ok(layout)
     }
