+++ core/embed/rust/src/ui/ui_firmware.rs
@@ -303,7 +303,7 @@ pub trait FirmwareUI {
 
     fn show_device_menu(
         failed_backup: bool,
-        battery_percentage: usize,
+        battery_percentage: u8,
         paired_devices: Vec<TString<'static>, 1>,
     ) -> Result<impl LayoutMaybeTrace, Error>;
